+++ python/lsst/ts/m2gui/__init__.py
@@ -29,9 +29,17 @@ else:
     except ImportError:
         __version__ = "?"
 
-from .enums import LocalMode
-from .utils import set_button, get_rgb_from_hex, colors
+from .enums import LocalMode, LimitSwitchType, Ring
+from .utils import set_button
+from .signals import (
+    SignalControl,
+    SignalMessage,
+    SignalError,
+    SignalStatus,
+    SignalLimitSwitch,
+)
 from .log_window_handler import LogWindowHandler
+from .fault_manager import FaultManager
 from .model import Model
 from .control_tabs import ControlTabs
 from .main_window import MainWindow