+++ python/lsst/ts/m2gui/controltab/__init__.py
@@ -29,4 +29,3 @@ from .tab_diagnostics import TabDiagnostics
 from .tab_overview import TabOverview
 from .tab_rigid_body_pos import TabRigidBodyPos
 from .tab_utility_view import TabUtilityView
-from .tab_limit_switch_status import TabLimitSwitchStatus