+++ python/lsst/ts/m2gui/model.py
@@ -21,8 +21,8 @@
 
 __all__ = ["Model"]
 
-from .enums import LocalMode
-from .signal import SignalError
+from .enums import LocalMode, LimitSwitchType, Ring
+from . import SignalStatus, FaultManager
 
 
 class Model(object):
@@ -39,12 +39,16 @@ class Model(object):
         A logger.
     is_csc_commander : `bool`
         Commandable SAL component (CSC) is the commander or not.
-    local_mode : `LocalMode`
+    local_mode : enum `LocalMode`
         Local model.
     is_closed_loop : `bool`
         Closed-loop control is on or not.
-    signal_error : `SignalError`
-        Signal to report the new or cleared errors.
+    signal_status : `SignalStatus`
+        Signal to report the updated status of system.
+    system_status : `dict`
+        System status.
+    fault_manager : `FaultManager`
+        Fault manager to record the system error.
     """
 
     def __init__(self, log):
@@ -55,19 +59,170 @@ class Model(object):
         self.local_mode = LocalMode.Standby
         self.is_closed_loop = False
 
-        self.signal_error = SignalError()
+        self.signal_status = SignalStatus()
+
+        self.system_status = self._set_system_status()
+
+        self.fault_manager = FaultManager(self.get_actuator_default_status(False))
+
+    def _set_system_status(self):
+        """Set the default system status.
+
+        Returns
+        -------
+        system_status : `dict`
+            Default system status.
+        """
+
+        system_status = {
+            "isCrioConnected": False,
+            "isTelemetryActive": False,
+            "isInPosition": False,
+            "isInMotion": False,
+            "isTcsActive": False,
+            "isPowerCommunicationOn": False,
+            "isPowerMotorOn": False,
+            "isOpenLoopMaxLimitsEnabled": False,
+            "isAlarmWarningOn": False,
+        }
+
+        return system_status
+
+    def get_actuator_default_status(self, status):
+        """Get the default actuator status.
+
+        Parameters
+        ----------
+        status : `any`
+            Default status.
+
+        Returns
+        -------
+        collection : `dict`
+            Collection of default actuator status.
+        """
+
+        collection = dict()
+        collection = self._set_actuator_default_status_specific_ring(
+            Ring.B, 30, collection, status
+        )
+        collection = self._set_actuator_default_status_specific_ring(
+            Ring.C, 24, collection, status
+        )
+        collection = self._set_actuator_default_status_specific_ring(
+            Ring.D, 18, collection, status
+        )
+        collection = self._set_actuator_default_status_specific_ring(
+            Ring.A, 6, collection, status
+        )
+
+        return collection
+
+    def _set_actuator_default_status_specific_ring(
+        self, ring, number, collection, status
+    ):
+        """Set the default actuator status in specific ring.
+
+        Parameters
+        ----------
+        ring : enum `Ring`
+            Name of ring.
+        number : `int`
+            Number of actuator in the ring.
+        collection : `dict`
+            Collection of actuator status.
+        status : `any`
+            Default status.
+
+        Returns
+        -------
+        collection : `dict`
+            Updated collection.
+        """
+
+        for idx in range(1, number + 1):
+            name = ring.name + str(idx)
+            collection[name] = status
+
+        return collection
+
+    def add_error(self, error):
+        """Add the error.
+
+        Parameters
+        ----------
+        error : `int`
+            Error code.
+        """
+
+        self.fault_manager.add_error(error)
+        self._check_error_and_update_status()
+
+    def _check_error_and_update_status(self):
+        """Check whenever the error is triggered, and update the related
+        internal status."""
+
+        is_error_on = self.fault_manager.has_error()
+        self.update_system_status("isAlarmWarningOn", is_error_on)
+
+    def clear_error(self, error):
+        """Clear the error.
+
+        Parameters
+        ----------
+        error : `int`
+            Error code.
+        """
+
+        self.fault_manager.clear_error(error)
+        self._check_error_and_update_status()
 
     def reset_errors(self):
         """Reset errors."""
 
         self.log.info("Reset all errors.")
 
+        self.fault_manager.reset_errors()
+        self.fault_manager.reset_limit_switch_status(LimitSwitchType.Retract)
+        self.fault_manager.reset_limit_switch_status(LimitSwitchType.Extend)
+
+        self._check_error_and_update_status()
+
     def enable_open_loop_max_limit(self):
         """Enable the maximum limit in open-loop control."""
 
         if (self.local_mode == LocalMode.Enable) and not self.is_closed_loop:
-            self.log.info("Enable the maximum limit in open-loop control.")
+            self.update_system_status("isOpenLoopMaxLimitsEnabled", True)
         else:
             self.log.info(
                 "Failed to enable the maximum limit. Only allow in Enabled state and open-loop control."
+            )
+
+    def disable_open_loop_max_limit(self):
+        """Disable the maximum limit in open-loop control."""
+        self.update_system_status("isOpenLoopMaxLimitsEnabled", False)
+
+    def update_system_status(self, status_name, new_status):
+        """Update the system status.
+
+        Parameters
+        ----------
+        status_name : `str`
+            Status name. It must be in the keys of self.system_status.
+        new_status : `bool`
+            New status.
+
+        Raises
+        ------
+        ValueError
+            Status name is not in the list.
+        """
+
+        if status_name in self.system_status.keys():
+            if self.system_status[status_name] != new_status:
+                self.system_status[status_name] = new_status
+                self.signal_status.name_status.emit((status_name, new_status))
+        else:
+            raise ValueError(
+                f"{status_name} not in the {list(self.system_status.keys())}."
             )