+++ python/lsst/ts/m2gui/utils.py
@@ -19,37 +19,37 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-__all__ = ["set_button", "get_rgb_from_hex", "colors"]
+__all__ = ["set_button"]
 
 from functools import partial
+
+from PySide2.QtCore import Qt
 from PySide2.QtWidgets import QPushButton
-from PySide2.QtGui import QColor
-
-colors = {
-    "Red": "#FF0000",
-    "Green": "#00FF00",
-    "Blue": "#0000FF",
-    "Black": "#000000",
-    "White": "#FFFFFF",
-    "Electric Green": "#41CD52",
-    "Dark Blue": "#222840",
-    "Yellow": "#F9E56d",
-}
+from PySide2.QtGui import QPalette
 
 
-def set_button(name, callback, *args, is_checkable=False):
+def set_button(
+    name, callback, *args, is_checkable=False, is_indicator=False, is_adjust_size=False
+):
     """Set the button.
 
     Parameters
     ----------
     name : `str`
         Button name.
-    callback : `func`
-        Callback function object to use in future partial calls.
+    callback : `func` or None
+        Callback function object to use in future partial calls. Put None
+        if you do not want to have the callback function connected.
     *args : `args`
-        Tuple of callback arguments to future partial calls.
+        Tuple of callback arguments to future partial calls. This is only
+        meaningful if the callback is not None.
     is_checkable : `bool`, optional
         The button is checkable or not. (the default is False)
+    is_indicator : `bool`, optional
+        The button is indicator or not. Put the callback to be None if the
+        value is True. (the default is False)
+    is_adjust_size : `bool`, optional
+        Adjust the size of button or not. (the default is False)
 
     Returns
     -------
@@ -61,26 +61,19 @@ def set_button(name, callback, *args, is_checkable=False):
     if is_checkable:
         button.setCheckable(is_checkable)
 
-    button.clicked.connect(partial(callback, *args))
+    if callback is not None:
+        button.clicked.connect(partial(callback, *args))
+
+    if is_indicator:
+        button.setEnabled(False)
+        button.setAutoFillBackground(True)
+
+        palette = button.palette()
+        palette.setColor(QPalette.ButtonText, Qt.black)
+
+        button.setPalette(palette)
+
+    if is_adjust_size:
+        button.adjustSize()
 
     return button
-
-
-def get_rgb_from_hex(code):
-    """Get RGB color from hexadecimal code.
-
-    Parameters
-    ----------
-    code : `str`
-        Hexadecimal code.
-
-    Returns
-    -------
-    `PySide2.QtGui.QColor`
-        Color.
-    """
-
-    code_hex = code.replace("#", "")
-    rgb = tuple(int(code_hex[idx : idx + 2], 16) for idx in (0, 2, 4))
-
-    return QColor.fromRgb(rgb[0], rgb[1], rgb[2])